requirements/development.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.4
Django==5.0
django-autoslug==1.9.9
django-cleanup==8.0.0
django-filter==23.5
django-rest-framework==0.1.0
django-silk==5.0.4
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
gprof2dot==2022.7.29
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Pillow==10.1.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
uritemplate==4.1.1
